refactor: Move package of StringCalculator from calculator to expression and rename to ExpressionCalculator
